Bulk inspection fix: redundant type cast
SVN r43850 |2016-05-06 18:06:16 +0000

+++ hdrl/src/org/labkey/hdrl/query/HDRLQuerySchema.java
@@ -107,7 +107,7 @@ public class HDRLQuerySchema extends SimpleUserSchema
                 statusMap.put(key, value);
             }
         }
-        return statusMap.get((Integer)s);
+        return statusMap.get(s);
     }
 
     public static void register(final HDRLModule module)
